Fix size of ssize_t

+++ libc/unistd/read.c
@@ -5,6 +5,6 @@
 
 DEFN_SYSCALL3(read,  SYS_READ, int, char *, size_t);
 
-int read(int file, void *ptr, size_t len) {
+ssize_t read(int file, void *ptr, size_t len) {
 	__sets_errno(syscall_read(file,ptr,len));
 }
